fix blinking
